pyproject.toml
[build-system]
requires = ["setuptools", "setuptools_scm"]
build-backend = "setuptools.build_meta"

[project]
name = "yutipy"
dynamic = ["version"]
dependencies = [
  "pykakasi==2.3.0",
  "python-dotenv==1.0.1",
  "rapidfuzz==3.12.1",
  "requests==2.32.3",
  "ytmusicapi==1.10.1",
]
requires-python = ">= 3.8"
authors = [
  {name = "Cheap Nightbot"},
  {name = "Cheap Nightbot", email = "hi@cheapnightbot.slmail.me"},
]
maintainers = [
  {name = "Cheap Nightbot", email = "hi@cheapnightbot.slmail.me"}
]
description = "A simple package for retrieving music information from various music platforms APIs."
readme = "README.md"
license = {file = "LICENSE"}
keywords = ["music", "API", "Deezer", "iTunes", "KKBox", "Spotify", "YouTube Music", "search", "retrieve", "information", "yutify"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "Topic :: Software Development :: Libraries",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]

[project.optional-dependencies]
dev =  [
    "pytest",
]

[project.urls]
Homepage = "https://github.com/CheapNightbot/yutipy"
Documentation = "https://yutipy.readthedocs.io/"
Repository = "https://github.com/CheapNightbot/yutipy.git"
Issues = "https://github.com/CheapNightbot/yutipy/issues"
Changelog = "https://github.com/CheapNightbot/yutipy/blob/master/CHANGELOG.md"
funding = "https://ko-fi.com/cheapnightbot"

[tool.setuptools_scm]
version_scheme = "guess-next-dev"
local_scheme = "no-local-version"

[project.scripts]
yutipy-cli = "cli.search:main"
yutipy-config = "cli.config:run_config_wizard"
